api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
torch
torchvision
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
